configure: fix handling of non linux operating systems
The OS whitelist conditional is malformed due to missing test(1) commands
causing the following error to be emitted during configure while using something
other than Linux:

	../configure[3340]: xopenbsd: not found

Instead, favor a switch statement which makes the line a bit shorter as well.

Also, the AC_MSG_ERROR argument must be wrapped in brackets due to presence of a
comma. Otherwise, the tail will be interpreted as an additional argument.